Explicit cast to (uint8_t)
We already ensured that
    0 <= slen <= MAXSLEN (:= 256 - sizeof(struct pnode))
so the cast is safe.

+++ lib/datastruct/patricia.c
@@ -322,7 +322,7 @@ patricia_insert(PATRICIA * P, const uint8_t * key, size_t keylen, void * rec)
 				slen = MAXSLEN;
 
 			/* Create the node or error out. */
-			if ((pnew = node_alloc(slen, key)) == NULL)
+			if ((pnew = node_alloc((uint8_t)slen, key)) == NULL)
 				return (-1);
 
 			/* Add the new node into the tree. */
